requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
joblib>=1.1.0
flasc>=0.1.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
pandas>=1.3.0
tqdm>=4.62.0
Pillow>=8.3.0
hdbscan>=0.8.2
